Server/requirements.txt
Flask==3.1.1
Flask_Cors==4.0.1
mysql_connector_repackaged==0.3.1
Werkzeug==3.1.3
